Merge develop in master

+++ src/main/java/com/denka88/ateliergrace/impl/MaterialServiceImpl.java
@@ -1,7 +1,9 @@
 package com.denka88.ateliergrace.impl;
 
 import com.denka88.ateliergrace.model.Material;
+import com.denka88.ateliergrace.model.Order;
 import com.denka88.ateliergrace.repo.MaterialRepo;
+import com.denka88.ateliergrace.repo.OrderRepo;
 import com.denka88.ateliergrace.service.MaterialService;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
@@ -13,9 +15,11 @@ import java.util.Optional;
 public class MaterialServiceImpl implements MaterialService {
     
     private final MaterialRepo materialRepo;
+    private final OrderRepo orderRepo;
 
-    public MaterialServiceImpl(MaterialRepo materialRepo) {
+    public MaterialServiceImpl(MaterialRepo materialRepo, OrderRepo orderRepo) {
         this.materialRepo = materialRepo;
+        this.orderRepo = orderRepo;
     }
 
     @Override
@@ -34,7 +38,22 @@ public class MaterialServiceImpl implements MaterialService {
     }
 
     @Override
+    @Transactional
     public void delete(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID материала не может быть null");
+        }
+        
+        Material material = materialRepo.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Материал с ID " + id + " не найден"));
+        
+        List<Order> orders = orderRepo.findByMaterialsId(id);
+
+        for (Order order : orders) {
+            order.getMaterials().remove(material);
+            orderRepo.save(order);
+        }
+
         materialRepo.deleteById(id);
     }
